requirements.txt
langchain-community==0.3.29
langchain-text-splitters==0.3.11
langgraph==0.6.6
langchain-google-genai==2.1.10
langchain-openai==0.3.32
langchain-huggingface==0.3.1
langchain-chroma==0.2.5
sentence-transformers==5.1.0
dotenv==0.9.9
playwright==1.55.0
unstructured==0.18.14
pdfminer.six==20250506
pi_heif==1.1.0
unstructured_inference==1.0.5
pdf2image==1.17.0
unstructured_pytesseract==0.3.15
pytesseract==0.3.13
streamlit==1.49.1
torch==